requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
pdf2image
Pillow
pyodbc
pyyaml